+++ HISTORY.md
@@ -1,3 +1,14 @@
+## [Unreleased]
+
+## [0.2.2] - 2024-01-25
+
+### Added
+
+- encode for set
+- encode for indexmap
+- encode references to traits within collections
+- provides a method for conversion to a string
+
 ## [0.2.1] - 2024-01-20
 
 ### Changed